refactor: SearchModal 폴더 분리
